Added the payer of and their coverage to the output files for Immunizations, Procedures, Medications. Added the number of Covered/Uncovered instances of each of the following to the payers.csv output file: Procedures, Immunizations, Medications.
